Update Rock paper scissors
In this code, the user is against the computer by pressing the button options. Other aspects such as the score limit to determine who the winner is and restart functions will be added later on.
The OOP version of the code will be uploaded once the basics are fully completed as expected.